split report from reportgen
