+++ inference_perf/reportgen/__init__.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .base import ReportGenerator, ReportFile
+from .base import ReportGenerator
 
 
-__all__ = ["ReportGenerator", "ReportGenerator", "ReportFile"]
+__all__ = ["ReportGenerator", "ReportGenerator"]
